Update Item view
 o Displays the list of items with Name, Category & Price

o Each item has 2 links: „Details“ and „Add To Cart“

o „Add To Cart“ link adds an item to the currently active Shopping Cart (the one with
Paid == false) and redirects to the details page of this Shopping Cart. If there’s no
active cart available, a new shopping cart is created.

o „Details“ link redirects to the item details page where Available Sizes are shown (in
addition to Name, Category &amp; Price). There’s an „Add To Cart“ link on the details
page.

o It’s not possible to create a new item or edit existing items from UI (they are seeded
to DB)