Update Shoppingcart view
o Displays the list of shopping carts with Sum, Date Created &amp; Paid, sorted by
DateCreated (last created on top).

o Each shopping cart has 2 links: Details and Delete

o „Details“ page displays (in addition to Sum and Date Created) the list of items in cart.
There’s a payment link on the details page. When user click this link, the value of.
Paid property is set to „true“ and user is redirected to the list of shopping carts.

o It’s not possible to create a new cart or edit existing carts directly from UI.